requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2
gunicorn==23.0.0
h11==0.16.0
packaging==24.2
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
